Per-currency amount precision: NUMERIC(14,2) money + decimals by currency
- service_records.cost/parts_cost -> NUMERIC(14,2) (exact decimal, lossless
  migration from int); Go money fields -> *float64; stats totals float64;
  passport/prompt sums updated.
- frontend money.js: per-currency decimals (RUB/UAH/KZT=0, BYN/USD/EUR=2);
  formatMoney renders with the right fraction digits; edit inputs use
  step 0.01 vs 1 and parseFloat based on the selected currency.

Verified: 10.50 + 150.99 = 161.49 round-trips; RUB shows no kopecks.

+++ apps/garagebrain/backend/internal/handler/stats.go
@@ -10,10 +10,10 @@ import (
 )
 
 type StatsResponse struct {
-	TotalCost     int            `json:"total_cost"`
-	RecordsByType map[string]int `json:"records_by_type"`
-	MonthlyCosts  map[string]int `json:"monthly_costs"`
-	RecordCount   int            `json:"record_count"`
+	TotalCost     float64            `json:"total_cost"`
+	RecordsByType map[string]float64 `json:"records_by_type"`
+	MonthlyCosts  map[string]float64 `json:"monthly_costs"`
+	RecordCount   int                `json:"record_count"`
 }
 
 func GetStats(w http.ResponseWriter, r *http.Request) {
@@ -35,13 +35,13 @@ func GetStats(w http.ResponseWriter, r *http.Request) {
 
 	stats := StatsResponse{
 		TotalCost:     0,
-		RecordsByType: make(map[string]int),
-		MonthlyCosts:  make(map[string]int),
+		RecordsByType: make(map[string]float64),
+		MonthlyCosts:  make(map[string]float64),
 		RecordCount:   len(records),
 	}
 
 	for _, r := range records {
-		amount := 0
+		amount := 0.0
 		if r.Cost != nil {
 			amount += *r.Cost
 		}
